# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Zanzibar Tourism API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Activity {
  id: number;
  name: string;
  description: string;
  /** e.g. water, cultural, adventure, nature, food */
  category: string;
  durationHours: number;
  pricePerPerson: number;
  imageUrl: string;
  isPopular: boolean;
}

export interface Accommodation {
  id: number;
  name: string;
  /** e.g. beach-resort, boutique-hotel, eco-lodge, luxury-villa, guesthouse */
  type: string;
  description: string;
  location: string;
  pricePerNight: number;
  imageUrl: string;
  amenities: string[];
  stars: number;
}

export interface Package {
  id: number;
  name: string;
  slug: string;
  tagline: string;
  description: string;
  /** e.g. beach, adventure, cultural, luxury, family */
  category: string;
  minDays: number;
  maxDays: number;
  basePricePerPersonPerDay: number;
  imageUrl: string;
  galleryImages: string[];
  includedActivities: Activity[];
  accommodation: Accommodation;
  highlights: string[];
  isFeatured: boolean;
  rating: number;
  reviewCount: number;
}

export interface EstimateRequest {
  /** @nullable */
  packageId?: number | null;
  accommodationId: number;
  activityIds: number[];
  numberOfVisitors: number;
  numberOfDays: number;
}

export interface CostBreakdownItem {
  label: string;
  amount: number;
  detail: string;
}

export interface EstimateResponse {
  accommodationCost: number;
  activitiesCost: number;
  guideCost: number;
  subtotal: number;
  taxes: number;
  totalCost: number;
  perPersonCost: number;
  breakdown: CostBreakdownItem[];
}

export interface CreateInquiryBody {
  firstName: string;
  lastName: string;
  email: string;
  phone?: string;
  /** @nullable */
  packageId?: number | null;
  numberOfVisitors: number;
  numberOfDays: number;
  preferredStartDate?: string;
  message: string;
  activityIds?: number[];
  /** @nullable */
  accommodationId?: number | null;
  /** @nullable */
  estimatedTotal?: number | null;
}

export interface Inquiry {
  id: number;
  firstName: string;
  lastName: string;
  email: string;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  packageId?: number | null;
  numberOfVisitors: number;
  numberOfDays: number;
  /** @nullable */
  preferredStartDate?: string | null;
  message: string;
  status: string;
  createdAt: string;
}

export interface Summary {
  totalPackages: number;
  totalActivities: number;
  totalAccommodations: number;
  totalInquiries: number;
  categories: string[];
  popularActivities: Activity[];
  featuredPackages: Package[];
}

export type ListPackagesParams = {
  category?: string;
  minDays?: number;
  maxDays?: number;
  maxPrice?: number;
};