# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Zanzibar Tourism API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /packages:
    get:
      operationId: listPackages
      tags: [packages]
      summary: List all tourism packages
      parameters:
        - name: category
          in: query
          required: false
          schema:
            type: string
        - name: minDays
          in: query
          required: false
          schema:
            type: integer
        - name: maxDays
          in: query
          required: false
          schema:
            type: integer
        - name: maxPrice
          in: query
          required: false
          schema:
            type: number
      responses:
        "200":
          description: List of packages
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Package"
  /packages/{id}:
    get:
      operationId: getPackage
      tags: [packages]
      summary: Get a single package
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Package details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Package"
        "404":
          description: Not found
...
components:
  schemas:
...
    Package:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        slug:
          type: string
        tagline:
          type: string
        description:
          type: string
        category:
          type: string
          description: "e.g. beach, adventure, cultural, luxury, family"
        minDays:
          type: integer
        maxDays:
          type: integer
        basePricePerPersonPerDay:
          type: number
        imageUrl:
          type: string
        galleryImages:
          type: array
          items:
            type: string
        includedActivities:
          type: array
          items:
            $ref: "#/components/schemas/Activity"
        accommodation:
          $ref: "#/components/schemas/Accommodation"
        highlights:
          type: array
          items:
            type: string
        isFeatured:
          type: boolean
        rating:
          type: number
        reviewCount:
          type: integer
      required:
        - id
        - name
        - slug
        - tagline
        - description
        - category
        - minDays
        - maxDays
        - basePricePerPersonPerDay
        - imageUrl
        - galleryImages
        - includedActivities
        - accommodation
        - highlights
        - isFeatured
        - rating
        - reviewCount
    Activity:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
        category:
          type: string
          description: "e.g. water, cultural, adventure, nature, food"
        durationHours:
          type: number
        pricePerPerson:
          type: number
        imageUrl:
          type: string
        isPopular:
          type: boolean
      required:
        - id
        - name
        - description
        - category
        - durationHours
        - pricePerPerson
        - imageUrl
        - isPopular
    Accommodation:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        type:
          type: string
          description: "e.g. beach-resort, boutique-hotel, eco-lodge, luxury-villa, guesthouse"
        description:
          type: string
        location:
          type: string
        pricePerNight:
          type: number
        imageUrl:
          type: string
        amenities:
          type: array
          items:
            type: string
        stars:
          type: integer
      required:
        - id
        - name
        - type
        - description
        - location
        - pricePerNight
        - imageUrl
        - amenities
        - stars
...