# lib/api-zod/src/generated/types/accommodation.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Zanzibar Tourism API
 * OpenAPI spec version: 0.1.0
 */

export interface Accommodation {
  id: number;
  name: string;
  /** e.g. beach-resort, boutique-hotel, eco-lodge, luxury-villa, guesthouse */
  type: string;
  description: string;
  location: string;
  pricePerNight: number;
  imageUrl: string;
  amenities: string[];
  stars: number;
}