# lib/api-zod/src/generated/types/activity.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Zanzibar Tourism API
 * OpenAPI spec version: 0.1.0
 */

export interface Activity {
  id: number;
  name: string;
  description: string;
  /** e.g. water, cultural, adventure, nature, food */
  category: string;
  durationHours: number;
  pricePerPerson: number;
  imageUrl: string;
  isPopular: boolean;
}