# lib/api-zod/src/generated/types/package.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Zanzibar Tourism API
 * OpenAPI spec version: 0.1.0
 */
import type { Accommodation } from "./accommodation";
import type { Activity } from "./activity";

export interface Package {
  id: number;
  name: string;
  slug: string;
  tagline: string;
  description: string;
  /** e.g. beach, adventure, cultural, luxury, family */
  category: string;
  minDays: number;
  maxDays: number;
  basePricePerPersonPerDay: number;
  imageUrl: string;
  galleryImages: string[];
  includedActivities: Activity[];
  accommodation: Accommodation;
  highlights: string[];
  isFeatured: boolean;
  rating: number;
  reviewCount: number;
}