# lib/db/src/schema/activities.ts
import { pgTable, text, serial, boolean, real } from "drizzle-orm/pg-core";
import { createInsertSchema } from "drizzle-zod";
import { z } from "zod/v4";

export const activitiesTable = pgTable("activities", {
  id: serial("id").primaryKey(),
  name: text("name").notNull(),
  description: text("description").notNull(),
  category: text("category").notNull(),
  durationHours: real("duration_hours").notNull(),
  pricePerPerson: real("price_per_person").notNull(),
  imageUrl: text("image_url").notNull(),
  isPopular: boolean("is_popular").notNull().default(false),
});

export const insertActivitySchema = createInsertSchema(activitiesTable).omit({ id: true });
export type InsertActivity = z.infer<typeof insertActivitySchema>;
export type Activity = typeof activitiesTable.$inferSelect;